Fixes for DTGE and NC
When an NC is used within a DTGE cell, the serialization of the nested value causes a problem during extraction.
Only during extraction, we should not serialize the JSON values, but return them as raw values.

The NC resolver needed to refactored to use `objects` rather than `JTokens`, it was causing type issues.